fix TSE comments
